+++ lib/omniauth_openid_federation/rack.rb
@@ -1,0 +1,2 @@
+require "rack"
+require_relative "rack_endpoint"