+++ lib/omniauth_openid_federation/user_info.rb
@@ -1,0 +1,15 @@
+module OmniauthOpenidFederation
+  class UserInfo
+    attr_reader :raw_attributes
+
+    def initialize(raw_attributes)
+      @raw_attributes = raw_attributes.each_with_object({}) do |(key, value), claims|
+        claims[key.to_sym] = value
+      end
+    end
+
+    def as_json(_options = {})
+      raw_attributes.transform_keys(&:to_s)
+    end
+  end
+end